Move update_config_with_includes from config to main
